Consolidate "queue once" functions - always use an idle timer
Helm is behaving fine with these now - seems that idle timers were being
suppressed in Helm prompts for another reason (never seen the behaviour before,
so it was strange).

This solves the issue where synchronization is delayed when opening a helm
prompt. When helm updates candidates, it seems to pause, meaning normal
timers run even if idle timers don't.

If this causes issues down the line, we can fix them later.