Fix yasnippet key sync definition - fire when snippets are changed
The yasnippet key sync definition was ported to the `voicemacs-define-sync`
macro incorrectly. It shouldn't fire on buffer change, it should fire when the
yasnippets are updated. The operation is expensive, so firing unnecessarily will
cause slowdown.

Fix this so yasnippets are synced only when snippets are updated. This fixes a
dramatic slowdown issue making Emacs snappy again.